Assume nullability of primitive types
At the end of the day, what we want here is null warning suppression, but doing it with an exclamation mark can be quite hidden, and can just result in exceptions sometime later at runtime. The AssumeNotNull method is an aggressive version that throws at the site of the actual null.

Also increased the language version to be able to use nameof to refer to a parameter.